Ship the latest version of stalld.
  - Shipping https://git.kernel.org/pub/scm/utils/stalld/stalld.git/
    c35b62722ffebb3cf16ad6b470d9275f007f6fa2
